+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - Launch configurations for Activity ([#20](https://github.com/KazeJiyu/ekumi/pull/20))
+- Automatic update of the History View ([#22](https://github.com/KazeJiyu/ekumi/pull/22))
 
 ## [0.1.0] - 2018-12-18
 ### Added